update docs and refactor
