+++ .travis.yml
@@ -1,10 +1,4 @@
 language: node_js
-
-cache:
-  directories:
-  - node_modules
-
 node_js:
-- 8.9.1
-
-script: npm run ci
+  - "8"
+after_success: ./node_modules/.bin/codecov
