+++ package.json
@@ -5,11 +5,10 @@
   "description": "Route versioning library.",
   "main": "index.js",
   "scripts": {
-    "test": "npm run test:coverage && npm run test:lint",
-    "test:coverage": "jest tests --coverage -i --forceExit",
-    "test:lint": "eslint src",
-    "docs": "jsdoc2md -t jsdoc2md/readme.hbs lib.js > readme.md",
-    "ci": "npm run test:lint && npm run test:coverage && codecov"
+    "test": "npm run test:lint && npm run test:coverage",
+    "test:coverage": "jest tests --coverage",
+    "test:lint": "eslint index",
+    "docs": "jsdoc2md -t jsdoc2md/readme.hbs index.js > readme.md"
   },
   "author": "Abdulrahman Amri",
   "license": "MIT",
